Move enhancer to renderer package
