+++ pkg/engine/task/task.go
@@ -4,27 +4,17 @@ import (
 	"errors"
 	"fmt"
 
+	"github.com/kudobuilder/kudo/pkg/engine/renderer"
+
 	"github.com/kudobuilder/kudo/pkg/apis/kudo/v1alpha1"
-	"github.com/kudobuilder/kudo/pkg/engine"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
-
-// Metadata contains Metadata along with specific fields associated with current plan
-// being executed like current plan, phase, step or task names.
-type Metadata struct {
-	engine.Metadata
-
-	PlanName  string
-	PhaseName string
-	StepName  string
-	TaskName  string
-}
 
 // Context is a engine.task execution context containing k8s client, templates parameters etc.
 type Context struct {
 	Client     client.Client
-	Enhancer   Enhancer
-	Meta       Metadata
+	Enhancer   renderer.Enhancer
+	Meta       renderer.Metadata
 	Templates  map[string]string // Raw templates
 	Parameters map[string]string // Instance and OperatorVersion parameters merged
 }
