+++ pkg/engine/task/task_apply_test.go
@@ -5,6 +5,8 @@ import (
 	"fmt"
 	"testing"
 
+	"github.com/kudobuilder/kudo/pkg/engine/renderer"
+
 	"github.com/kudobuilder/kudo/pkg/engine"
 	"github.com/kudobuilder/kudo/pkg/util/template"
 	"github.com/stretchr/testify/assert"
@@ -18,7 +20,7 @@ import (
 )
 
 func TestApplyTask_Run(t *testing.T) {
-	meta := Metadata{
+	meta := renderer.Metadata{
 		Metadata: engine.Metadata{
 			InstanceName:        "test",
 			InstanceNamespace:   "default",
@@ -51,7 +53,7 @@ func TestApplyTask_Run(t *testing.T) {
 			ctx: Context{
 				Client:   fake.NewFakeClientWithScheme(scheme.Scheme),
 				Enhancer: &testEnhancer{},
-				Meta:     Metadata{},
+				Meta:     renderer.Metadata{},
 			},
 		},
 		{
@@ -168,7 +170,7 @@ func resourceAsString(resource metav1.Object) string {
 
 type testEnhancer struct{}
 
-func (k *testEnhancer) Apply(templates map[string]string, metadata Metadata) ([]runtime.Object, error) {
+func (k *testEnhancer) Apply(templates map[string]string, metadata renderer.Metadata) ([]runtime.Object, error) {
 	result := make([]runtime.Object, 0)
 	for _, t := range templates {
 		objsToAdd, err := template.ParseKubernetesObjects(t)
@@ -182,6 +184,6 @@ func (k *testEnhancer) Apply(templates map[string]string, metadata Metadata) ([]
 
 type errorEnhancer struct{}
 
-func (k *errorEnhancer) Apply(templates map[string]string, metadata Metadata) ([]runtime.Object, error) {
+func (k *errorEnhancer) Apply(templates map[string]string, metadata renderer.Metadata) ([]runtime.Object, error) {
 	return nil, errors.New("always error")
 }
