+++ pkg/engine/workflow/engine_test.go
@@ -5,9 +5,10 @@ import (
 	"testing"
 	"time"
 
+	"github.com/kudobuilder/kudo/pkg/engine/renderer"
+
 	"github.com/kudobuilder/kudo/pkg/apis/kudo/v1alpha1"
 	"github.com/kudobuilder/kudo/pkg/engine"
-	"github.com/kudobuilder/kudo/pkg/engine/task"
 	"github.com/kudobuilder/kudo/pkg/util/template"
 	"github.com/pkg/errors"
 	corev1 "k8s.io/api/core/v1"
@@ -37,7 +38,7 @@ func TestExecutePlan(t *testing.T) {
 		metadata       *engine.Metadata
 		expectedStatus *v1alpha1.PlanStatus
 		wantErr        bool
-		enhancer       task.Enhancer
+		enhancer       renderer.Enhancer
 	}{
 		{name: "plan already finished will not change its status", activePlan: &ActivePlan{
 			Name: "test",
@@ -636,7 +637,7 @@ func instance() *v1alpha1.Instance {
 
 type testEnhancer struct{}
 
-func (k *testEnhancer) Apply(templates map[string]string, metadata task.Metadata) ([]runtime.Object, error) {
+func (k *testEnhancer) Apply(templates map[string]string, metadata renderer.Metadata) ([]runtime.Object, error) {
 	result := make([]runtime.Object, 0)
 	for _, t := range templates {
 		objsToAdd, err := template.ParseKubernetesObjects(t)
